refactor: Extract common test fakes to shared binding container
