+++ core/integration/infra/src/jvmMain/kotlin/com/thomaskioko/tvmaniac/testing/di/TestJvmPlatformBindingContainer.kt
@@ -7,16 +7,8 @@ import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
 import com.thomaskioko.tvmaniac.core.base.AppPreferencesDataStore
 import com.thomaskioko.tvmaniac.core.base.IoCoroutineScope
-import com.thomaskioko.tvmaniac.core.logger.CrashReporter
-import com.thomaskioko.tvmaniac.core.logger.fixture.FakeCrashReporter
-import com.thomaskioko.tvmaniac.core.notifications.api.NotificationManager
-import com.thomaskioko.tvmaniac.core.notifications.testing.FakeNotificationManager
-import com.thomaskioko.tvmaniac.core.tasks.api.BackgroundTaskScheduler
-import com.thomaskioko.tvmaniac.core.tasks.testing.FakeBackgroundTaskScheduler
 import com.thomaskioko.tvmaniac.datastore.implementation.DATA_STORE_FILE_NAME
 import com.thomaskioko.tvmaniac.db.TvManiacDatabase
-import com.thomaskioko.tvmaniac.util.api.FormatterUtil
-import com.thomaskioko.tvmaniac.util.testing.FakeFormatterUtil
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.BindingContainer
 import dev.zacsweers.metro.ContributesTo
@@ -53,20 +45,4 @@ public object TestJvmPlatformBindingContainer {
             produceFile = { tempDir.resolve(DATA_STORE_FILE_NAME).toString().toPath() },
         )
     }
-
-    @Provides
-    @SingleIn(AppScope::class)
-    public fun provideFormatterUtil(): FormatterUtil = FakeFormatterUtil()
-
-    @Provides
-    @SingleIn(AppScope::class)
-    public fun provideCrashReporter(): CrashReporter = FakeCrashReporter()
-
-    @Provides
-    @SingleIn(AppScope::class)
-    public fun provideNotificationManager(): NotificationManager = FakeNotificationManager()
-
-    @Provides
-    @SingleIn(AppScope::class)
-    public fun provideBackgroundTaskScheduler(): BackgroundTaskScheduler = FakeBackgroundTaskScheduler()
 }
